Remove ignored files

+++ HotelManagement/Models/Hotel.cs
@@ -5,6 +5,5 @@ public class Hotel
     public int Id { get; set; }
     public string Name { get; set; }
     public string Address { get; set; }
-    public int StarRating { get; set; }
-
+    public short StarRating { get; set; }
 }
